requirements.txt
altair==5.2.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
clarifai==9.11.0
clarifai-grpc==9.11.0
click==8.1.7
colorama==0.4.6
contextlib2==21.6.0
gitdb==4.0.11
GitPython==3.1.41
googleapis-common-protos==1.62.0
grpcio==1.60.0
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.3
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.3
opencv-python==4.7.0.68
packaging==23.2
pandas==2.1.4
pillow==10.2.0
protobuf==4.25.2
pyarrow==14.0.2
pydeck==0.8.1b0
Pygments==2.17.2
Pympler==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-rapidjson==1.14
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
rich==13.4.2
rpds-py==0.17.1
schema==0.7.5
six==1.16.0
smmap==5.0.1
streamlit==1.30.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
tritonclient==2.34.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==4.3.1
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
